Move audit serialiser to datahub.core
